food-review-analyzer/backend/requirements.txt
Flask
Flask-Cors
python-dotenv
SQLAlchemy
transformers
torch
google-generative-ai
